+++ schema_drift/integrations/__init__.py
@@ -1,4 +1,8 @@
 """schema_drift.integrations"""
-from .slack import notify as slack_notify, make_notifier as slack_notifier
+from .slack import notify as slack_notify, make_notifier as slack_notifier, SlackNotifier
+from .team import approve, annotate, audit_log, pending_approvals, require_approval
 
-__all__ = ["slack_notify", "slack_notifier"]
+__all__ = [
+    "slack_notify", "slack_notifier", "SlackNotifier",
+    "approve", "annotate", "audit_log", "pending_approvals", "require_approval",
+]